Remove extra delete listener

+++ pinlockview/src/main/java/com/andrognito/pinlockview/PinLockAdapter.java
@@ -371,6 +371,9 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
             mDeleteButton.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
+                    if (mDeleteButton.getVisibility() == View.GONE) {
+                        return;
+                    }
                     if (mOnDeleteClickListener != null) {
                         mOnDeleteClickListener.onDeleteClicked();
                     }
@@ -380,6 +383,9 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
             mDeleteButton.setOnLongClickListener(new View.OnLongClickListener() {
                 @Override
                 public boolean onLongClick(View v) {
+                    if (mDeleteButton.getVisibility() == View.GONE) {
+                        return true;
+                    }
                     if (mOnDeleteClickListener != null) {
                         mOnDeleteClickListener.onDeleteLongClicked();
                     }
@@ -392,6 +398,9 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
 
                 @Override
                 public boolean onTouch(View v, MotionEvent event) {
+                    if (mDeleteButton.getVisibility() == View.GONE) {
+                        return false;
+                    }
                     if (event.getAction() == MotionEvent.ACTION_DOWN) {
                         mButtonImage.setColorFilter(mCustomizationOptionsBundle
                                 .getDeleteButtonPressesColor());
@@ -435,6 +444,9 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
             mEnterButton.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
+                    if (mEnterButton.getVisibility() == View.GONE) {
+                        return;
+                    }
                     if (mOnEnterClickListener != null) {
                         mOnEnterClickListener.onEnterClicked();
                     }
@@ -446,6 +458,9 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
 
                 @Override
                 public boolean onTouch(View v, MotionEvent event) {
+                    if (mEnterButton.getVisibility() == View.GONE) {
+                        return false;
+                    }
                     if (event.getAction() == MotionEvent.ACTION_DOWN) {
                         mEnterButton.setColorFilter(mCustomizationOptionsBundle.getEnterButtonPressesColor(),
                                 PorterDuff.Mode.SRC_ATOP);
